refactor: extract image_service.go from ocr_vedem.go
